Updated comments, renamed constants, and improved code readability
In this commit, several changes were made across multiple files to improve code readability and maintainability. The comments in `BaseEntity.cs` and `Employee.cs` were translated from Portuguese to English to clarify the relationship between entities. No actual changes were made to `GeneralResponse.cs`, but a line was removed and then added back. The text associated with an icon in `NavMenu.razor` was changed for better understanding. A non-descriptive comment line was removed from `Program.cs`. The `BaseLibrary.DTOs` namespace was removed from `_Imports.razor` to clean up the import statements. In `CustomAuthenticationStateProvider.cs`, a comment was added to indicate the injection of `LocalStorageService` and a space was removed for code consistency. The constant `AuthUrl` was renamed to `AuThUrl` in `UserAccountService.cs` for better naming convention. Also, comments were added to several methods in `UserAccountService.cs`, `IUserAccount.cs`, and `UserAccountRepository.cs` to explain their purpose and steps.