Updated authentication and layout in various components
Updated several components in the application. In `MainLayout.razor`, removed several namespaces and updated the `LogoutClicked` method to use `CustomAuthenticationStateProvider`. Reformatted `NavMenu` and updated `AuthorizeView` to include a `context` parameter. In `LoginPage.razor` and `RegisterPage.razor`, removed `DialogService` and added a `CascadingParameter` for `AuthenticationState`. In `HomePage.razor`, changed the layout to `TableBanner` and added a `CascadingParameter` for `AuthenticationState`. In `Program.cs`, replaced `HttpClient` registration with a scoped service registration and updated the comment for `Syncfusion` dialog service registration. Created a new component `TableBanner.razor` with a grid of cards representing different categories.