`Update NavMenu behavior and clean up MainLayout`
In this commit, we've made a couple of key changes to the `NavMenu.razor` and `MainLayout.razor` files.

In `NavMenu.razor`, we've updated the initial value of `collapseNavMenu` from `true` to `false`, meaning the navigation menu will now be expanded by default. We've also modified the `ToggleNavMenu()` method so that it no longer toggles the value of `collapseNavMenu`. Instead, it now sets `collapseNavMenu` to its own value, which effectively makes the method do nothing.

In `MainLayout.razor`, we've made some non-functional changes. The `<main>` tag and its contents, as well as the `LogoutClicked()` method in the `@code` section, were removed and then added back in the same position. This indicates that there were no actual changes made to these parts of the code.