Refactor NavMenu.razor and update navigation links
Significant refactoring has been done on the `NavMenu.razor` file. The greeting message in the navbar has been updated from "Hi" to "Welcome". The `NavLink` for the Home page now points to `/home/dashboard` instead of the root directory. A typo in the `NavLink` for the Counter page has been corrected. A new `NavLink` for an Administration section has been added with a sub-link for Users. The `ToggleNavMenu` function and `collapseNavMenu` variable have been removed. A `TODO` comment has been added for future work.